Add service type
