+++ src/Features/Core/Portable/InitializeParameter/IInitializeParameterService.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.CodeAnalysis.Editing;
+using Microsoft.CodeAnalysis.Host;
+using Microsoft.CodeAnalysis.Operations;
+
+namespace Microsoft.CodeAnalysis.InitializeParameter;
+
+internal interface IInitializeParameterService : ILanguageService
+{
+    void AddAssignment(
+        SyntaxNode constructorDeclaration,
+        IBlockOperation? blockStatement,
+        IParameterSymbol parameter,
+        ISymbol fieldOrProperty,
+        SyntaxEditor editor);
+}
